refactor(cli): consolidate internal command-tree wiring into internal.go
